reorganized code a bit and cleaned up tests
